Upgrade to SpotBugs 3.1.3
This replaces FindBugs.  Also address a few violations.

+++ compute/src/main/java/org/jclouds/compute/util/AutomaticHardwareIdSpec.java
@@ -75,7 +75,7 @@ public class AutomaticHardwareIdSpec {
    @Override
    public String toString() {
       if (disk.isPresent()) {
-         return String.format("automatic:cores=%s;ram=%s;disk=%.0f", cores, ram, disk.get().floatValue());
+         return String.format("automatic:cores=%s;ram=%s;disk=%.0f", cores, ram, disk.get());
       }
       else {
          return String.format("automatic:cores=%s;ram=%s", cores, ram);
